requirements.txt
python-dotenv
PIL